# ky-admin/swagger/swagger.yaml
basePath: /api
definitions:
  api.LoginRequest:
    properties:
      password:
        description: 密码
        example: password123
        type: string
      username:
        description: 用户名
        example: admin
        type: string
    required:
    - password
    - username
    type: object
  api.LoginResponse:
    properties:
      token:
        description: JWT令牌
        type: string
      userId:
        description: 用户ID
        type: integer
    type: object
  api.RefreshTokenRequest:
    properties:
      refreshToken:
        description: 刷新令牌
        type: string
    required:
    - refreshToken
    type: object
  api.RefreshTokenResponse:
    properties:
      refreshToken:
        description: 新刷新令牌
        type: string
      token:
        description: 新访问令牌
        type: string
    type: object
  api.Response:
    properties:
      code:
        description: 状态码
        type: integer
      data:
        description: 数据
      message:
        description: 消息
        type: string
      page:
        description: 当前页（用于分页）
        type: integer
      size:
        description: 每页大小（用于分页）
        type: integer
      total:
        description: 总数（用于分页）
        type: integer
    type: object
  api.TokenResponse:
    properties:
      refreshToken:
        description: 刷新令牌
        type: string
      token:
        description: 访问令牌
        type: string
      user:
        description: 用户信息
    type: object
  dto.ChangePasswordReq:
    properties:
      newPassword:
        type: string
      oldPassword:
        type: string
    required:
    - newPassword
    - oldPassword
    type: object
  model.LoginDTO:
    properties:
      password:
        description: 密码
        example: password123
        type: string
      username:
        description: 用户名
        example: admin
        type: string
    required:
    - password
    - username
    type: object
  model.RegisterDTO:
    properties:
      email:
        description: 邮箱
        example: admin@example.com
        type: string
      nickname:
        description: 昵称
        example: 管理员
        type: string
      password:
        description: 密码
        example: password123
        type: string
      phone:
        description: 电话
        example: "13800138000"
        type: string
      roleIds:
        description: 角色ID列表
        items:
          type: integer
        type: array
      username:
        description: 用户名
        example: admin
        type: string
    required:
    - password
    - username
    type: object
  model.RoleDTO:
    properties:
      code:
        type: string
      createdAt:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      permissionIds:
        items:
          type: integer
        type: array
      sort:
        type: integer
      status:
        type: integer
      tenantId:
        type: integer
      updatedAt:
        type: string
    type: object
  model.UpdateUserDTO:
    properties:
      avatar:
        description: 头像
        type: string
      email:
        description: 邮箱
        example: admin@example.com
        type: string
      nickname:
        description: 昵称
        example: 管理员
        type: string
      phone:
        description: 电话
        example: "13800138000"
        type: string
      roleIds:
        description: 角色ID列表
        items:
          type: integer
        type: array
      status:
        description: 状态
        example: 1
        type: integer
    type: object
  model.UserDTO:
    properties:
      avatar:
        description: 头像
        type: string
      createdAt:
        description: 创建时间
        type: string
      deptId:
        description: 部门ID
        type: integer
      email:
        description: 邮箱
        type: string
      id:
        description: 用户ID
        type: integer
      nickname:
        description: 昵称
        type: string
      phone:
        description: 电话
        type: string
      roles:
        description: 角色列表
        items:
          $ref: '#/definitions/model.RoleDTO'
        type: array
      status:
        description: 状态
        type: integer
      tenantId:
        description: 租户ID
        type: integer
      updatedAt:
        description: 更新时间
        type: string
      username:
        description: 用户名
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@kysion.com
    name: API Support
    url: http://www.kysion.com/support
  description: KY-Admin后台管理系统API文档
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: KY-Admin API
  version: "1.0"
paths:
  /api/auth/login:
    post:
      consumes:
      - application/json
      description: 用户登录接口
      parameters:
      - description: 登录信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/api.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 登录成功
          schema:
            allOf:
            - $ref: '#/definitions/api.Response'
            - properties:
                data:
                  $ref: '#/definitions/api.LoginResponse'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/api.Response'
        "401":
          description: 认证失败
          schema:
            $ref: '#/definitions/api.Response'
      summary: 用户登录
      tags:
      - 用户管理
  /api/auth/refresh:
    post:
      consumes:
      - application/json
      description: 刷新用户访问令牌
      parameters:
      - description: 刷新令牌信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/api.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 刷新成功
          schema:
            allOf:
            - $ref: '#/definitions/api.Response'
            - properties:
                data:
                  $ref: '#/definitions/api.RefreshTokenResponse'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/api.Response'
        "401":
          description: 令牌无效
          schema:
            $ref: '#/definitions/api.Response'
      summary: 刷新令牌
      tags:
      - 认证管理
  /api/auth/register:
    post:
      consumes:
      - application/json
      description: 用户注册接口
      parameters:
      - description: 注册信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.RegisterDTO'
      produces:
      - application/json
      responses:
        "200":
          description: 注册成功
          schema:
            allOf:
            - $ref: '#/definitions/api.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.UserDTO'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/api.Response'
      summary: 用户注册
      tags:
      - 认证管理
  /api/user/info:
    get:
      consumes:
      - application/json
      description: 获取当前登录用户的详细信息
      produces:
      - application/json
      responses:
        "200":
          description: 用户信息
          schema:
            allOf:
            - $ref: '#/definitions/api.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.UserDTO'
              type: object
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/api.Response'
      security:
      - Bearer: []
      summary: 获取当前用户信息
      tags:
      - 用户管理
    put:
      consumes:
      - application/json
      description: 更新当前登录用户的信息
      parameters:
      - description: 用户信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.UpdateUserDTO'
      produces:
      - application/json
      responses:
        "200":
          description: 更新后的用户信息
          schema:
            allOf:
            - $ref: '#/definitions/api.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.UserDTO'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/api.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/api.Response'
      security:
      - Bearer: []
      summary: 更新用户信息
      tags:
      - 用户管理
  /api/user/password:
    put:
      consumes:
      - application/json
      description: 修改当前登录用户的密码
      parameters:
      - description: 密码信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.ChangePasswordReq'
      produces:
      - application/json
      responses:
        "200":
          description: 密码修改成功
          schema:
            $ref: '#/definitions/api.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/api.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/api.Response'
      security:
      - Bearer: []
      summary: 修改密码
      tags:
      - 用户管理
  /api/users:
    get:
      consumes:
      - application/json
      description: 分页获取用户列表
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: pageSize
        type: integer
      - description: 用户名
        in: query
        name: username
        type: string
      - description: 昵称
        in: query
        name: nickname
        type: string
      - description: 状态
        in: query
        name: status
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 用户列表
          schema:
            allOf:
            - $ref: '#/definitions/api.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.UserDTO'
                  type: array
              type: object
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/api.Response'
      security:
      - Bearer: []
      summary: 获取用户列表
      tags:
      - 用户管理
  /api/users/{id}:
    get:
      consumes:
      - application/json
      description: 根据用户ID获取用户详细信息
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 用户信息
          schema:
            allOf:
            - $ref: '#/definitions/api.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.UserDTO'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/api.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/api.Response'
        "404":
          description: 用户不存在
          schema:
            $ref: '#/definitions/api.Response'
      security:
      - Bearer: []
      summary: 根据ID获取用户信息
      tags:
      - 用户管理
  /health:
    get:
      consumes:
      - application/json
      description: 返回服务健康状态
      produces:
      - application/json
      responses:
        "200":
          description: 服务健康
          schema:
            allOf:
            - $ref: '#/definitions/api.Response'
            - properties:
                data:
                  additionalProperties:
                    type: string
                  type: object
              type: object
      summary: 健康检查接口
      tags:
      - 健康检查
  /livez:
    get:
      consumes:
      - application/json
      description: 用于Kubernetes存活探测
      produces:
      - application/json
      responses:
        "200":
          description: 服务存活
          schema:
            allOf:
            - $ref: '#/definitions/api.Response'
            - properties:
                data:
                  additionalProperties:
                    type: string
                  type: object
              type: object
      summary: 存活探针
      tags:
      - 健康检查
  /readyz:
    get:
      consumes:
      - application/json
      description: 用于Kubernetes就绪探测
      produces:
      - application/json
      responses:
        "200":
          description: 服务就绪
          schema:
            allOf:
            - $ref: '#/definitions/api.Response'
            - properties:
                data:
                  additionalProperties:
                    type: string
                  type: object
              type: object
      summary: 就绪探针
      tags:
      - 健康检查
schemes:
- http
- https
swagger: "2.0"